splitted threads functions in folders and files matching the type of ship
